(*.[ch]):	patches to types to wipe warnings from compilers 	with more strict noise levels. (also good when compiling 	with gcc -Wall -pedantic -Wwrite-strings -Wcast-qual)
git-svn-id: file:///home/hardaker/lib/sf-bkups/net-snmp-convert-svnrepo/trunk/net-snmp@2584 06827809-a52a-0410-b366-d66718629ded

+++ agent/mibgroup/host_res.h
@@ -67,7 +67,7 @@ typedef	void (*PFV) (void);
 typedef	int  (*PFI) (int);
 typedef	int  (*PFIV) (void);
 typedef	const char* (*PFS) (int);
-typedef	oid* (*PFO) (int, int *);
+typedef	oid* (*PFO) (int, size_t *);
 
 extern PFV init_device[];	/* Routines for stepping through devices */
 extern PFIV next_device[];
